add new context and component folder refactor, adds external link component
