+++ example.env.bt
@@ -78,6 +78,9 @@ DISPLAY_DEBUG_MENU=false
 DISPLAY_COVID_DATA=false
 STATE_ABBREVIATION=
 COVID_ACT_NOW_API_KEY=
+EXTERNAL_COVID_DATA_LINK=https://google.com
+EXTERNAL_COVID_DATA_LABEL=Covid Data Label
+EXTERNAL_TRAVEL_GUIDENCE_LINK="https://maps.google.com
 
 // Age Verification
 DISPLAY_AGE_VERIFICATION=true
