+++ src/EscrowVerification/MoreInfo.tsx
@@ -12,7 +12,7 @@ import { useCustomCopy } from "../configuration/useCustomCopy"
 import { SvgXml } from "react-native-svg"
 
 import { Text } from "../components"
-import { useConfigurationContext } from "../ConfigurationContext"
+import { useConfigurationContext } from "../configuration"
 
 import { Spacing, Typography, Colors, Buttons } from "../styles"
 import { Icons } from "../assets"
