+++ src/Settings/ProductAnalyticsConsentScreen.tsx
@@ -9,7 +9,7 @@ import {
 import { useNavigation } from "@react-navigation/native"
 import { useTranslation } from "react-i18next"
 
-import { useConfigurationContext } from "../ConfigurationContext"
+import { useConfigurationContext } from "../configuration"
 import { useProductAnalyticsContext } from "../ProductAnalytics/Context"
 import { useStatusBarEffect } from "../navigation"
 import { useCustomCopy } from "../configuration/useCustomCopy"
