+++ src/components/ListItemSeparator/index.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from "react"
 import { View, StyleSheet } from "react-native"
 
-import { Outlines } from "../styles"
+import { Outlines } from "../../styles"
 
 const ListItemSeparator: FunctionComponent = () => {
   return <View style={style.separator} />
@@ -13,4 +13,4 @@ const style = StyleSheet.create({
   },
 })
 
-export default ListItemSeparator
+export { ListItemSeparator }
