--- src/components/index.ts
@@ -1,15 +1,0 @@
-import Text from "./Text"
-import StatusBar from "./StatusBar"
-import ListItem from "./ListItem"
-import ListItemSeparator from "./ListItemSeparator"
-import LoadingIndicator from "./LoadingIndicator"
-import SwitchListItem from "./SwitchListItem"
-
-export {
-  Text,
-  StatusBar,
-  ListItem,
-  ListItemSeparator,
-  LoadingIndicator,
-  SwitchListItem,
-}
