+++ src/factories/index.ts
@@ -1,5 +1,5 @@
 import affectedUserFlowContext from "./affectedUserFlowContext"
-import configurationContext from "./configurationContext"
+import configurationContext from "./configuration"
 import covidDataContext from "./covidDataContext"
 import exposureContext from "./exposureContext"
 import exposureDatum from "./exposureDatum"
