+++ src/navigation/MainTabNavigator.tsx
@@ -12,7 +12,7 @@ import ExposureHistoryStack from "./ExposureHistoryStack"
 import HomeStack from "./HomeStack"
 import SymptomHistoryStack from "./SymptomHistoryStack"
 import SettingsStack from "./SettingsStack"
-import { useConfigurationContext } from "../ConfigurationContext"
+import { useConfigurationContext } from "../configuration"
 import { TabRoute, TabRoutes } from "./index"
 import { TabBarIcons } from "../assets/svgs/TabBarNav"
 import { Text } from "../components"
